requirements.txt
beautifulhue=0.1.2
RPi.GPIO=0.6.5
adafruit-circuitpython-ads1x15=1.0.2
python-dotenv=0.10.1